+++ sn_transfers/src/cashnotes/address.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::UniquePubkey;
+use super::UniquePubkey;
 
 use serde::{Deserialize, Serialize};
 use std::hash::Hash;
