+++ sn_transfers/src/cashnotes/fee_output.rs
