+++ sn_transfers/src/transfers/mod.rs
@@ -27,50 +27,8 @@
 //! sending of tokens from one address to another. Or many.
 //! A cash_note transaction is the lower layer concept where the blinded inputs and outputs are specified.
 
-mod error;
+mod offline_transfer;
 mod transfer;
 
-use std::collections::BTreeMap;
-
-use xor_name::XorName;
-
-use crate::{CashNote, SignedSpend, Transaction, UniquePubkey};
-
-pub use self::error::{Error, Result};
-pub type ContentPaymentsIdMap = BTreeMap<XorName, Vec<UniquePubkey>>;
-
-/// Utility function to create an offline transfer
-pub use self::transfer::create_offline_transfer;
-
-/// Offline Transfer
-/// This struct contains all the necessary information to carry out the transfer.
-/// The created cash_notes and change cash_note from a transfer
-/// of tokens from one or more cash_notes, into one or more new cash_notes.
-#[derive(custom_debug::Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq)]
-pub struct OfflineTransfer {
-    /// This is the transaction where all the below
-    /// spends were made and cash_notes created.
-    pub tx: Transaction,
-    /// The cash_notes that were created containing
-    /// the tokens sent to respective recipient.
-    #[debug(skip)]
-    pub created_cash_notes: Vec<CashNote>,
-    /// The cash_note holding surplus tokens after
-    /// spending the necessary input cash_notes.
-    #[debug(skip)]
-    pub change_cash_note: Option<CashNote>,
-    /// The parameters necessary to send all spend requests to the network.
-    pub all_spend_requests: Vec<SpendRequest>,
-}
-
-/// The parameters necessary to send a spend request to the network.
-#[derive(
-    custom_debug::Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq, Eq, PartialOrd, Ord,
-)]
-pub struct SpendRequest {
-    /// The cash_note to register in the network as spent.
-    pub signed_spend: SignedSpend,
-    /// The cash_note transaction that the spent cash_note was created in.
-    #[debug(skip)]
-    pub parent_tx: Transaction,
-}
+pub use offline_transfer::{create_offline_transfer, ContentPaymentsIdMap, OfflineTransfer};
+pub use transfer::{CashNoteRedemption, Transfer};
